Move EvHandler::event to Layout
