+++ src/traits.rs
@@ -10,10 +10,12 @@ use std::ops::DerefMut;
 use std::time::Duration;
 
 use crate::draw::{DrawHandle, SizeHandle};
-use crate::event::{self, Manager, ManagerState};
+use crate::event::{self, Event, Manager, ManagerState, Response};
 use crate::geom::{Coord, Rect, Size};
 use crate::layout::{self, AxisInfo, SizeRules};
 use crate::{AlignHints, CoreData, WidgetId};
+
+mod impls;
 
 /// Support trait for cloning boxed unsized objects
 #[cfg_attr(not(feature = "internal_doc"), doc(hidden))]
@@ -200,11 +202,11 @@ pub trait Widget: WidgetCore {
 
 /// Positioning and drawing routines for widgets
 ///
-/// This trait contains methods concerned with positioning of contents, other
-/// than those in [`event::Handler`].
+/// This trait contains methods concerned with positioning of contents
+/// as well as low-level event handling.
 ///
 /// For a description of the widget size model, see [`SizeRules`].
-pub trait Layout: event::EvHandler {
+pub trait Layout: event::Handler {
     /// Get size rules for the given axis.
     ///
     /// This method takes `&mut self` to allow local caching of child widget
@@ -262,6 +264,34 @@ pub trait Layout: event::EvHandler {
     /// This method is called to draw each visible widget (and should not
     /// attempt recursion on child widgets).
     fn draw(&self, draw_handle: &mut dyn DrawHandle, mgr: &ManagerState);
+
+    /// Handle a low-level event.
+    ///
+    /// Most non-parent widgets will not need to implement this method manually.
+    /// The default implementation (which wraps [`Manager::handle_generic`])
+    /// forwards high-level events via [`Handler::action`], thus the only reason
+    /// for non-parent widgets to implement this manually is for low-level
+    /// event processing.
+    ///
+    /// Parent widgets should forward events to the appropriate child widget,
+    /// via logic like the following:
+    /// ```
+    /// if id <= self.child1.id() {
+    ///     self.child1.event(mgr, id, event).into()
+    /// } else if id <= self.child2.id() {
+    ///     self.child2.event(mgr, id, event).into()
+    /// } else {
+    ///     debug_assert!(id == self.id(), "Layout::event: bad WidgetId");
+    ///     // either handle `event`, or return:
+    ///     Response::Unhandled(event)
+    /// }
+    /// ```
+    /// Optionally, the return value of child event handlers may be intercepted
+    /// in order to handle returned messages and/or unhandled events.
+    #[inline]
+    fn event(&mut self, mgr: &mut Manager, _: WidgetId, event: Event) -> Response<Self::Msg> {
+        Manager::handle_generic(self, mgr, event)
+    }
 }
 
 /// Trait to describe the type needed by the layout implementation.
