+++ src/widget/scrollbar.rs
@@ -182,6 +182,10 @@ impl<D: Directional> ScrollBar<D> {
     }
 }
 
+impl<D: Directional> event::Handler for ScrollBar<D> {
+    type Msg = u32;
+}
+
 impl<D: Directional> Layout for ScrollBar<D> {
     fn size_rules(&mut self, size_handle: &mut dyn SizeHandle, axis: AxisInfo) -> SizeRules {
         let (size, min_len) = size_handle.scrollbar();
@@ -214,13 +218,7 @@ impl<D: Directional> Layout for ScrollBar<D> {
         let hl = mgr.highlight_state(self.handle.id());
         draw_handle.scrollbar(self.core.rect, self.handle.rect(), dir, hl);
     }
-}
 
-impl<D: Directional> event::Handler for ScrollBar<D> {
-    type Msg = u32;
-}
-
-impl<D: Directional> event::EvHandler for ScrollBar<D> {
     fn event(&mut self, mgr: &mut Manager, id: WidgetId, event: Event) -> Response<Self::Msg> {
         let offset = if id <= self.handle.id() {
             match self.handle.event(mgr, id, event).try_into() {
